requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
redis
django-redis
drf-spectacular
djangorestframework-simplejwt
django-oauth-toolkit
django-debug-toolbar
django-extensions
whitenoise
pillow
boto3
django-storages
celery
django-celery-beat
django-celery-results
channels
channels-redis
django-ratelimit
python-decouple
gunicorn
dj-database-url
django-allauth
dj-rest-auth
djangorestframework-simplejwt
pyotp
qrcode
twilio
moviepy
imageio
imageio-ffmpeg
stripe
pillow
celery[redis]
django-redis
livekit
livekit-api
daphne
drf-nested-routers
livekit-protocol
requests
PyJWT